Fix #22: Only update gh-pages when we on the master branch

+++ travis.sh
@@ -4,14 +4,18 @@ set -e -o pipefail
 
 dub test --compiler=${DC}
 
-if [ ! -z "$GH_TOKEN" ]; then
-    dub build -b docs --compiler=${DC}
-    cd docs
-    git init
-    git config user.name "Travis-CI"
-    git config user.email "travis@nodemeatspace.com"
-    git add .
-    git commit -m "Deployed to Github Pages"
-    git push --force --quiet "https://${GH_TOKEN}@github.com/${TRAVIS_REPO_SLUG}" HEAD:gh-pages > /dev/null 2>&1
-    #git push --force "https://${GH_TOKEN}@github.com/${TRAVIS_REPO_SLUG}" HEAD:gh-pages
+branch=$(git branch | sed -n -e 's/^\* \(.*\)/\1/p')
+
+if [ "$branch" == "master" ]; then
+    if [ ! -z "$GH_TOKEN" ]; then
+        dub build -b docs --compiler=${DC}
+        cd docs
+        git init
+        git config user.name "Travis-CI"
+        git config user.email "travis@nodemeatspace.com"
+        git add .
+        git commit -m "Deployed to Github Pages"
+        git push --force --quiet "https://${GH_TOKEN}@github.com/${TRAVIS_REPO_SLUG}" HEAD:gh-pages > /dev/null 2>&1
+        #git push --force "https://${GH_TOKEN}@github.com/${TRAVIS_REPO_SLUG}" HEAD:gh-pages
+    fi
 fi
